FILE: moves groupIngredientsByType closer to usage site
